+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Tidarr notable changes.
 - [Front] Add toggle finished items button in queue list
 - [Docker] Load custom `/config/requirements.txt` file to extend your scripts
 - [API] Update Tiddl to 3.2.3
+- [API] Allow POST queries over Tidal proxy
 
 ## 📦 1.2.0
 ### 🐛 Fixed